djangoapp/requirements.txt
django>=5.1,<5.2
psycopg2-binary>=2.9.9,<2.10
requests>=2.32.3,<2.33
xmltodict>=0.13.0,<0.14
# psycopg2>=2.9.9,<2.10
pandas>=2.2.2,<2.3
djangorestframework